requirements.txt
discord.py
flask
python-dotenv
re
asyncio
aiohttp
time
io